Fix: prDBT and syspth arrays — uint32_t[] not ulong[]
Each entry holds a 4-byte big-endian OS-9 value; ulong[] gave 8-byte stride
which would produce a malformed table for any OS-9 code walking these arrays.
fcalls.c D_PrcDBT ptr narrowed to uint32_t* to match.

+++ Source/OS9exec_core/fcalls.c
@@ -999,9 +999,9 @@ os9err OS9_F_SetSys( regs_type *rp, ushort cpid )
     ulong        b   = TO68K(mdirField);
     process_typ* cp  = &procs[cpid];
 
-    ulong  v;
-    ulong* ptr;
-    int    k;
+    ulong     v;
+    uint32_t* ptr;
+    int       k;
         
     switch (offs) {
       case D_ID      : v=                   MODSYNC; break;
